+++ packages/e2e-tests/src/selectors.ts
@@ -131,9 +131,24 @@ export class RegisterUsernameModal {
     return this.driver.wait(until.elementLocated(By.xpath("//h3[text()='Register a username']")))
   }
 
+  get error() {
+    return this.driver.wait(until.elementLocated(By.xpath("//p[text()='Username already taken.']")))
+  }
+
   async typeUsername(username: string) {
     const usernameInput = await this.driver.findElement(By.xpath('//input[@name="userName"]'))
     await usernameInput.sendKeys(username)
+  }
+
+  async clearInput() {
+    const usernameInput = await this.driver.findElement(By.xpath('//input[@name="userName"]'))
+    if (process.platform === 'darwin') {
+      await usernameInput.sendKeys(Key.COMMAND + 'a')
+      await usernameInput.sendKeys(Key.DELETE)
+    } else {
+      await usernameInput.sendKeys(Key.CONTROL + 'a')
+      await usernameInput.sendKeys(Key.DELETE)
+    }
   }
 
   async submit() {