+++ packages/mobile/src/components/ConnectionProcess/ConnectionProcess.types.ts
@@ -1,0 +1,4 @@
+export interface ConnectionProcessComponentProps {
+  connectionProcess: { number: number; text: string }
+  openUrl: (url: string) => void
+}