+++ packages/mobile/src/store/nativeServices/events/nativeEvent.keys.ts
@@ -1,4 +1,7 @@
 export enum NativeEventKeys {
   Backend = 'backend',
   Notification = 'notification',
+  Stop = 'stop',
+  AppPause = 'apppause',
+  AppResume = 'appresume',
 }